Fix MSVC warnings in benchmark code

+++ test/benchmarks/benchmark_adler32.cc
@@ -42,7 +42,7 @@ public:
         uint32_t hash = 0;
 
         for (auto _ : state) {
-            hash = adler32(hash, (const unsigned char *)random_ints, state.range(0));
+            hash = adler32(hash, (const unsigned char *)random_ints, (size_t)state.range(0));
         }
 
         benchmark::DoNotOptimize(hash);
